pyproject.toml
[tool.poetry]
name = "protonmailgenerator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
PyAutoGUI = "^0.9.54"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
